clusterresolver: move tests around to different files
